sadadem
upgrade android gradle plugin
upgrade kotklin
add images
finish homepage
