+++ lib/main.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
-import 'homepage.dart';
+import 'home/homepage.dart';
 
 void main() {
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
