+++ pubspec.yaml
@@ -63,7 +63,9 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/img/stta.jpg
-  #   - images/a_dot_ham.jpeg
+    - assets/img/demografi.png
+    - assets/img/ekonomi.png
+    - assets/img/lingkungan.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
